Move resample_to_rate and fill_holes to new files under src/time_series/pipeline/
